[util] Move win32 specific implementations to their own files
