+++ src/util/platform/util_string_win32.cpp
