+++ src/util/util_env.cpp
@@ -1,7 +1,5 @@
 #include "util_env.h"
 
-#include "./com/com_include.h"
-
 namespace dxvk::env {
 
   std::string getEnvVar(const char* name) {
@@ -19,38 +17,6 @@ namespace dxvk::env {
     return (n != std::string::npos)
       ? fullPath.substr(n + 1)
       : fullPath;
-  }
-
-
-  std::string getExePath() {
-    std::vector<WCHAR> exePath;
-    exePath.resize(MAX_PATH + 1);
-
-    DWORD len = ::GetModuleFileNameW(NULL, exePath.data(), MAX_PATH);
-    exePath.resize(len);
-
-    return str::fromws(exePath.data());
-  }
-  
-  
-  void setThreadName(const std::string& name) {
-    using SetThreadDescriptionProc = HRESULT (WINAPI *) (HANDLE, PCWSTR);
-
-    static auto proc = reinterpret_cast<SetThreadDescriptionProc>(
-      ::GetProcAddress(::GetModuleHandleW(L"kernel32.dll"), "SetThreadDescription"));
-
-    if (proc != nullptr) {
-      auto wideName = std::vector<WCHAR>(name.length() + 1);
-      str::tows(name.c_str(), wideName.data(), wideName.size());
-      (*proc)(::GetCurrentThread(), wideName.data());
-    }
-  }
-
-
-  bool createDirectory(const std::string& path) {
-    WCHAR widePath[MAX_PATH];
-    str::tows(path.c_str(), widePath);
-    return !!CreateDirectoryW(widePath, nullptr);
   }
   
 }
